requirements.txt
telethon==1.36.0
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
